Reorganize frontend folder
