+++ frontend/src/App.tsx
@@ -2,14 +2,14 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import { CssBaseline } from "@mui/material";
 import React, { useState, useReducer, useEffect } from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Welcome from "./components/pages/Welcome";
-import Signup from "./components/pages/Signup";
+import Welcome from "./components/auth/WelcomePage";
+import Signup from "./components/auth/SignupPage";
 import PrivateRoute from "./components/auth/PrivateRoute";
 import Default from "./components/pages/Default";
 import CreateModulePage from "./components/pages/CreateModulePage";
 import NotFound from "./components/pages/NotFound";
 import NotAuthorized from "./components/pages/NotAuthorized";
-import MyAccount from "./components/pages/MyAccount";
+import MyAccount from "./components/pages/MyAccountPage";
 import AUTHENTICATED_USER_KEY from "./constants/AuthConstants";
 import AuthContext from "./contexts/AuthContext";
 import { getLocalStorageObj } from "./utils/LocalStorageUtils";
@@ -27,9 +27,9 @@ import ManageUserPage from "./components/pages/ManageUserPage";
 import MakeHelpRequestPage from "./components/pages/MakeHelpRequestPage";
 import ViewHelpRequestsPage from "./components/pages/ViewHelpRequestsPage";
 import HelpRequestPage from "./components/pages/HelpRequestPage";
-import CreatePasswordPage from "./components/pages/CreatePasswordPage";
+import CreatePasswordPage from "./components/auth/CreatePasswordPage";
 import ForgotPasswordPage from "./components/auth/forgot_password/ForgotPasswordPage";
-import CourseUnitsPage from "./components/pages/courses/CourseUnitsPage";
+import CourseViewingPage from "./components/course_viewing/CourseViewingPage";
 
 const App = (): React.ReactElement => {
   const currentUser: AuthenticatedUser | null =
@@ -133,7 +133,7 @@ const App = (): React.ReactElement => {
                 <PrivateRoute
                   exact
                   path={Routes.COURSE_PAGE}
-                  component={CourseUnitsPage}
+                  component={CourseViewingPage}
                   allowedRoles={["Administrator", "Facilitator", "Learner"]}
                 />
                 <Route exact path="*" component={NotFound} />
